+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[FIX]** Fixed `ItemCheckbox` a11y focus ring display for keyboard navigation
+- **[FIX]** Fixed `ItemEditableInfo` a11y keyboard navigation by Providing a focus ring 
 [...]
 
 # v40.0.0 (24/08/2020)